# access_control_service/access-control-service-openapi.yaml
openapi: 3.0.3
info:
  title: Compliance & Access Control Service API
  description: |
    Микросервис для управления геозонами, правилами доступа и контроля соблюдения политик безопасности.
    Обеспечивает привязку меток к сущностям (сотрудники/оборудование), проверку разрешенных зон и генерацию оповещений о нарушениях.
  version: 1.0.0
  contact:
    name: Команда разработки
    email: security@rtls.example.com
servers:
  - url: http://localhost:8084/api/v1
    description: Локальный сервер разработки
  - url: https://compliance.rtls.example.com/api/v1
    description: Продуктивный сервер
...
paths:
...
  /rules:
    get:
      tags:
        - Rules
      summary: Получение списка всех правил
      description: Возвращает список всех правил доступа и контроля.
      operationId: getAllRules
      parameters:
        - name: is_active
          in: query
          required: false
          schema:
            type: boolean
          description: Фильтр по активности правил
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'

    post:
      tags:
        - Rules
      summary: Создание нового правила
      description: Создает новое правило доступа или контроля.
      operationId: createRule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleCreate'
            examples:
              timeBasedRule:
                summary: Правило с временными ограничениями
                value:
                  name: "Доступ в серверную только в рабочее время"
                  description: "Сотрудники могут входить в серверную только с 9:00 до 18:00"
                  entity_type: "employee"
                  geofence_id: "550e8400-e29b-41d4-a716-446655440000"
                  action: "allow"
                  schedule:
                    days_of_week: [1, 2, 3, 4, 5]
                    start_time: "09:00"
                    end_time: "18:00"
                  severity: "medium"
              roleBasedRule:
                summary: Правило на основе ролей
                value:
                  name: "Только инженеры к оборудованию"
                  description: "Только сотрудники с ролью engineer могут приближаться к опасному оборудованию"
                  entity_type: "employee"
                  role_required: "engineer"
                  geofence_id: "550e8400-e29b-41d4-a716-446655440001"
                  action: "allow"
                  severity: "high"
      responses:
        '201':
          description: Правило успешно создано.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '400':
          description: Некорректные данные правила.
  /rules/{rule_id}:
    get:
      tags:
        - Rules
      summary: Получение информации о правиле
      description: Возвращает детальную информацию о правиле.
      operationId: getRuleById
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Успешный запрос.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Правило с указанным ID не найдено.

    patch:
      tags:
        - Rules
      summary: Обновление правила
      description: Обновляет данные правила (например, активность).
      operationId: updateRule
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RuleUpdate'
      responses:
        '200':
          description: Правило успешно обновлено.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          description: Правило с указанным ID не найдено.

    delete:
      tags:
        - Rules
      summary: Удаление правила
      description: Удаляет правило из системы.
      operationId: deleteRule
      parameters:
        - name: rule_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Правило успешно удалено.
        '404':
          description: Правило с указанным ID не найдено.

  # ==================== Compliance: Проверка соблюдения ====================
...
components:
  schemas:
...
    RuleCreate:
      type: object
      required:
        - name
        - entity_type
        - geofence_id
        - action
      properties:
        name:
          type: string
          description: Название правила
        description:
          type: string
          description: Описание правила
        entity_type:
          type: string
          enum: [employee, equipment, all]
          description: Тип сущности, к которой применяется правило
        entity_id:
          type: string
          description: Конкретная сущность (если null - применяется ко всем)
        role_required:
          type: string
          description: Требуемая роль (только для сотрудников)
        geofence_id:
          type: string
          format: uuid
          description: ID геозоны, к которой применяется правило
        action:
          type: string
          enum: [allow, deny, alert]
          description: Действие правила
        schedule:
          type: object
          description: Расписание действия правила
          properties:
            days_of_week:
              type: array
              items:
                type: integer
                minimum: 0
                maximum: 6
              description: Дни недели (0-воскресенье, 6-суббота)
            start_time:
              type: string
              pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
              description: Время начала (HH:MM)
            end_time:
              type: string
              pattern: '^([0-1][0-9]|2[0-3]):[0-5][0-9]$'
              description: Время окончания (HH:MM)
        severity:
          type: string
          enum: [low, medium, high, critical]
          default: medium
          description: Серьезность нарушения
        is_active:
          type: boolean
          default: true
        metadata:
          type: object
    RuleUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        is_active:
          type: boolean
        severity:
          type: string
          enum: [low, medium, high, critical]
        metadata:
          type: object
    Rule:
      allOf:
        - $ref: '#/components/schemas/RuleCreate'
        - type: object
          properties:
            rule_id:
              type: string
              format: uuid
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time

    # --- Проверка соблюдения ---
...